// src/models/fleet.model.ts
export interface Fleet {
  AllianceCountryCode: string;
  AllianceDescription: string;
  AllianceId: string;
  AllianceName: string;
  AllianceShipUserId: string;
  AllianceSpriteId: string;
  ChampionshipScore: string;
  ChannelId: string;
  Credits: string;
  DivisionDesignId: string;
  EnableWars: string;
  ImmunityDate: string;
  MinTrophyRequired: string;
  NumberOfApprovedMembers: string;
  NumberOfMembers: string;
  Ranking: string;
  RequiresApproval: string;
  Score: string;
  Trophy: string;
}