// src/models/user.model.ts
export interface User {
  ActivatedPromotions: string;
  AllianceId: string;
  AllianceJoinDate: string;
  AllianceMembership: string;
  AllianceName: string;
  AllianceQualifyDivisionDesignId: string;
  AllianceScore: string;
  AllianceSpriteId: string;
  AllianceSupplyDonation: string;
  AuthenticationType: string;
  BlockAuthAttemptsUntilDate: string;
  BoostAmount: string;
  BoostEndDate: string;
  CaptainCharacterDesignId: string;
  ChallengeDesignId: string;
  ChallengeLosses: string;
  ChallengeWins: string;
  ChampionshipScore: string;
  ChatAppearance: string;
  CompletedMissionDesigns: string;
  CompletedMissionEventIds: string;
  CooldownExpiry: string;
  CreationDate: string;
  CrewDonated: string;
  CrewReceived: string;
  DailyChallengeWinStreak: string;
  DailyMissionsAttempted: string;
  DailyRewardStatus: string;
  DrawsUsedToday: string;
  ExploredStarSystemIds: string;
  FacebookToken: string;
  FacebookTokenExpiryDate: string;
  Flags: string;
  FreeStarbuxReceivedToday: string;
  GameCenterFriendCount: string;
  GameCenterName: string;
  GenderType: string;
  GooglePlayIdToken: string;
  GooglePlayName: string;
  GoolePlayAuthCode: string;
  HeroBonusChance: string;
  HighestTrophy: string;
  IconSpriteId: string;
  Id: string;
  LanguageKey: string;
  LastAlertDate: string;
  LastBoostDate: string;
  LastCatalogPurchaseDate: string;
  LastChallengeDesignId: string;
  LastHeartBeatDate: string;
  LastLoginDate: string;
  LastPurchaseDate: string;
  LastRewardActionDate: string;
  LastVipClaimDate: string;
  LoadingPercentage: string;
  Name: string;
  PVPAttackDraws: string;
  PVPAttackLosses: string;
  PVPAttackWins: string;
  PVPDefenceDraws: string;
  PVPDefenceLosses: string;
  PVPDefenceWins: string;
  ProfileImageUrl: string;
  PurchaseRewardPoints: string;
  RaceType: string;
  Ranking: string;
  ShipDesignId: string;
  SituationOccurencesToday: string;
  Status: string;
  TipStatus: string;
  TotalSupplyDonation: string;
  TournamentBonusScore: string;
  TournamentRewardPoints: string;
  Trophy: string;
  TutorialStatus: string;
  UnlockedCharacterDesignIds: string;
  UnlockedShipDesignIds: string;
  UsedRewardPoints: string;
  UserType: string;
  VipExpiryDate: string;
}